+++ Tests/SentryTests/Integrations/Session/SentrySessionTrackerTests.swift
@@ -60,25 +60,25 @@ class SentrySessionTrackerTests: XCTestCase {
         super.setUp()
         
         fixture = Fixture()
-        
+
         SentryDependencyContainer.sharedInstance().dateProvider = fixture.currentDateProvider
 
         fixture.fileManager.deleteCurrentSession()
         fixture.fileManager.deleteCrashedSession()
         fixture.fileManager.deleteTimestampLastInForeground()
-        
+
         fixture.setNewHubToSDK()
-        
+
         sut = fixture.getSut()
     }
     
     override func tearDown() {
-        stopSut()
+        abnormalStopSut()
         clearTestState()
-        
+
         super.tearDown()
     }
-    
+
     func testOnlyForeground() throws {
         // -- Arrange --
         startSutInAppDelegate()
@@ -196,8 +196,18 @@ class SentrySessionTrackerTests: XCTestCase {
     
     func testCrashInBackground_LaunchInForeground() throws {
         // -- Arrange --
-        crashInBackground()
-        assertNoSessionSent()
+        // During testing we observed that the deallocation of the session tracker happens after the method returns
+        // and not immediately when a new `sut` is set.
+        // This causes multiple session tracker to be registered as observers, until they are fully released and
+        // weak references are nil.
+        // Using an autoreleasepool to ensure that the deallocation happens before the test continues.
+        autoreleasepool {
+            crashInBackground()
+            assertNoSessionSent()
+
+            // Manually deallocate the previous sut to avoid race-conditions of duplicate observers
+            sut = nil
+        }
 
         // -- Act & Assert --
         sut = fixture.getSut()
@@ -229,7 +239,7 @@ class SentrySessionTrackerTests: XCTestCase {
         goToBackground(forSeconds: 2)
         advanceTime(bySeconds: 2)
         // This is not a crash but an abnormal end.
-        stopSut()
+        abnormalStopSut()
 
         advanceTime(bySeconds: 1)
 
@@ -251,7 +261,7 @@ class SentrySessionTrackerTests: XCTestCase {
         advanceTime(bySeconds: 5)
         goToBackground()
         // This is not a crash but an abnormal end.
-        stopSut()
+        abnormalStopSut()
 
         advanceTime(bySeconds: 1)
 
@@ -343,26 +353,37 @@ class SentrySessionTrackerTests: XCTestCase {
         // -- Assert --
         assertEndSessionSent(started: sessionStarted, duration: 1)
     }
-    
+
     func testForeground_Background_Terminate_LaunchAgain() throws {
+        // During testing we observed that the deallocation of the session tracker happens after the method returns
+        // and not immediately when a new `sut` is set.
+        // This causes multiple session tracker to be registered as observers, until they are fully released and
+        // weak references are nil.
+        // Using an autoreleasepool to ensure that the deallocation happens before the test continues.
         // -- Arrange --
         let sessionStartTime = fixture.currentDateProvider.date()
         startSutInAppDelegate()
 
         // -- Act --
-        goToForeground()
-        advanceTime(bySeconds: 1)
-        goToBackground()
-        
-        advanceTime(bySeconds: 10)
-        terminateApp()
-        assertEndSessionSent(started: sessionStartTime, duration: 1)
-        stopSut()
+        autoreleasepool {
+            goToForeground()
+            advanceTime(bySeconds: 1)
+            goToBackground()
 
-        advanceTime(bySeconds: 1)
-        
-        // Launch the app again
-        fixture.setNewHubToSDK()
+            advanceTime(bySeconds: 10)
+            terminateApp()
+            assertEndSessionSent(started: sessionStartTime, duration: 1)
+            abnormalStopSut()
+
+            advanceTime(bySeconds: 1)
+
+            // Launch the app again
+            fixture.setNewHubToSDK()
+
+            // Manually deallocate the previous sut to avoid race-conditions of duplicate observers
+            sut = nil
+        }
+
         sut = fixture.getSut()
 
         startSutInAppDelegate()
@@ -438,6 +459,57 @@ class SentrySessionTrackerTests: XCTestCase {
         assertNotificationNames(notificationNames)
     }
 
+    func testForegroundBeforeStart_shoudStartSession() throws {
+        // -- Arrange --
+        goToForeground()
+
+        // Pre-condition: No session should be sent yet
+        assertNoSessionSent()
+
+        // -- Act --
+        sut.start()
+
+        // -- Assert --
+        XCTAssertEqual(fixture.client.captureSessionInvocations.invocations.count, 1)
+        let startSessionEvent = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 0))
+        XCTAssertEqual(startSessionEvent.status, SentrySessionStatus.ok)
+    }
+
+    func testRestartInForeground_shouldStartNewSession() throws {
+        // -- Arrange --
+        startSutInAppDelegate()
+        goToForeground()
+
+        XCTAssertEqual(fixture.client.captureSessionInvocations.invocations.count, 1)
+        var event = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 0))
+        XCTAssertEqual(event.status, SentrySessionStatus.ok)
+
+        stopSut()
+        // Ideally we would expect the SDK to send the `exited` session status when calling `stop`.
+        //
+        // As the test case `testKillAppWithoutNotificationsAndNoCrash_EndsWithAbnormalSession` is calling `stop` and
+        // expects that to result in an `abnormal` session next time the tracker is started, we have a contradiction.
+        //
+        // To keep existing tests working as intended, we decided to keep the `abnormal` status being send after the
+        // restart has happened.
+
+        // -- Act --
+        sut.start()
+
+        // -- Assert --
+        event = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 1))
+        XCTAssertEqual(event.status, SentrySessionStatus.abnormal)
+
+        event = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 2))
+        XCTAssertEqual(event.status, SentrySessionStatus.ok)
+
+        event = try XCTUnwrap(fixture.client.captureSessionInvocations.invocations.element(at: 3))
+        XCTAssertEqual(event.status, SentrySessionStatus.exited)
+
+        // Assert that there are no more invocations
+        XCTAssertEqual(fixture.client.captureSessionInvocations.invocations.count, 4)
+    }
+
     // MARK: - Helpers
 
     private func startSutInAppDelegate() {
@@ -453,6 +525,10 @@ class SentrySessionTrackerTests: XCTestCase {
     }
 
     private func stopSut() {
+        sut.stop()
+    }
+
+    private func abnormalStopSut() {
         sut.stop()
         #if os(iOS) || targetEnvironment(macCatalyst) || os(tvOS)
         // When the app stops, the app state is `inactive`.
@@ -481,7 +557,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // This can be observed by viewing the application state in `UIAppDelegate.applicationDidBecomeActive`.
         fixture.application.applicationState = .active
         #endif
-        Dynamic(sut).didBecomeActive()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.didBecomeActiveNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func goToBackground() {
@@ -490,8 +573,15 @@ class SentrySessionTrackerTests: XCTestCase {
         #if os(iOS) || targetEnvironment(macCatalyst) || os(tvOS)
         // It is expected that the app state is background when the didEnterBackground is called
         fixture.application.applicationState = .background
+        fixture.notificationCenter
+           .post(
+               Notification(
+                   name: SentryNSNotificationCenterWrapper.didEnterBackgroundNotificationName,
+                   object: nil,
+                   userInfo: nil
+               )
+           )
         #endif
-        Dynamic(sut).didEnterBackground()
     }
     
     private func willResignActive() {
@@ -500,7 +590,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // This can be observed by viewing the application state in `UIAppDelegate.applicationWillResignActive`.
         fixture.application.applicationState = .active
         #endif
-        Dynamic(sut).willResignActive()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.willResignActiveNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func hybridSdkDidBecomeActive() {
@@ -508,7 +605,14 @@ class SentrySessionTrackerTests: XCTestCase {
         // When an app did become active, it is in the active state.
         fixture.application.applicationState = .active
         #endif
-        Dynamic(sut).didBecomeActive()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.didBecomeActiveNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func goToBackground(forSeconds: TimeInterval) {
@@ -523,12 +627,19 @@ class SentrySessionTrackerTests: XCTestCase {
         #if os(iOS) || targetEnvironment(macCatalyst) || os(tvOS)
         fixture.application.applicationState = .background
         #endif
-        Dynamic(sut).willTerminate()
+        fixture.notificationCenter
+            .post(
+                Notification(
+                    name: SentryNSNotificationCenterWrapper.willTerminateNotificationName,
+                    object: nil,
+                    userInfo: nil
+                )
+            )
     }
     
     private func terminateApp() {
         willTerminate()
-        stopSut()
+        abnormalStopSut()
     }
     
     private func launchBackgroundTaskAppNotRunning() {
@@ -705,6 +816,10 @@ class SentrySessionTrackerTests: XCTestCase {
         override var applicationState: UIApplication.State {
             get { _underlyingAppState }
             set { _underlyingAppState = newValue }
+        }
+
+        override func isActive() -> Bool {
+            return applicationState == .active
         }
     }
 #endif